pyproject.toml
[tool.poetry]
name = "deep-diffusion"
version = "0.1.0"
description = ""
authors = ["Jan Latko s193223@student.dtu.dk", "Artur Przybysz s202384@student.dtu.dk"]

[tool.poetry.dependencies]
python = "^3.8"
wandb = "^0.12.1"
torch = "^1.9.0"
Pillow = "^8.3.2"
pytorch-lightning = "^1.4.6"
omegaconf = "^2.1.1"
hydra-core = "^1.1.1"
torchvision = "^0.10.0"
matplotlib = "^3.4.3"
tqdm = "^4.62.2"
pandas = "^1.3.3"
pytorch_ema = { git = "https://github.com/fadel/pytorch_ema", branch = "master" }


[tool.poetry.dev-dependencies]
poethepoet = "^0.10.0"
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
precommit = "python -m black scripts/ src/"
